Use an IPv6-enabled network in CLI's compose file
Add comments to the compose file that explain some decisions made there

Put Postgres on the same network with the sync service

Should've done that from the beginning. Without this, the sync service
cannot resolve the domain name of the postgres service when both are
used, i.e. when `electric-sql start --with-postgres` is invoked.